[FIX] mrp: always show post inventory button
The inventory button was shown or hide depending of the product valuation.
As the valuation was previously stored on the stock moves and posting
split the stock moves. Fifo and Avco valuation was not well computed.
The migration to stock valuation layer fixed this behavior. This
commit shows the post inventory button as soon as there are some
finished move to be posted.

task : 54665

closes odoo/odoo#41781

X-original-commit: eab4b576346412a06baa93aad9d8e9fc56d13f9f
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>