[FIX] point_of_sale: skip 0% tax declaration
Install point_of_sale and a localization (i.e. l10n_be) with a tax
configured with 0% and no account.
Assign the tax to a product available on POS. Open POS session, sell
such product, close session and validate. Error message will popup,
blocking the action because the tax used for the product has no
associated account for posting the entry.

This should not be the case because the tax is 0% so it should not be
posted. Skipping taxes based on the amount to avoid the issue.

opw-2155173

closes odoo/odoo#41821

X-original-commit: c7936d2884c10865bfc39a43d30aef6926f29079
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>