[FIX] stock: cast string as date
Writing for the web interface on the scheduled date of a stock move
will failed as the date is send to the python server as a string. Then,
a duration in datetime format is added. This failed as operation on string
and datetime is forbidden.
This commit cast the string date into a datetime object before applying
date logic.

closes odoo/odoo#42287

X-original-commit: 8f122297f80e06c06a3a5efc6815b9ba627d0ed8
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>