[FIX] project: fix sub tasks issue when callback through the breadcrum
Purpose of the task is that when creating a multi level sub-task
and going back on click of sub-task breadcrumb, sub task is not
going to visible. We need to refresh or go to the parent task
than after its visible.

so display all the sub task when click is press on sub-task
breadcrumb.
replaced action domain with 'child_of' instead of 'in' because the
child_of will return the same result as method returns.

closes odoo/odoo#42916

Task: 2157055
Closes: #42163
X-original-commit: 208c3352a8f4253a900a4c8985576b3fc0a7f76e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>